requirements.txt
scapy>=2.4.5
Click>=8.0
PyYAML>=5.4
pydantic>=1.8
jinja2>=3.0
pdfkit>=0.6
python-dotenv>=0.19
